Refactor header files and adjust .clang-format settings
This update includes a rearrangement of `#include` lines in the `include/resolver.hpp` file for better organization, and it eliminates the need for `include/mx_rdata.hpp`.
Changes were also made to the `include/dns_answer.hpp` file to correct the namespace.
Lastly, the `.clang-format` file was altered to set `MaxEmptyLinesToKeep` from 2 to 1 for maintaining more formatted code structure.