Replace asio with boost::asio and delete dns_response.cpp
The decision to replace the asio library with boost::asio in all files was taken to leverage the extra functionalities of boost::asio. Moreover, the dns_response.cpp file was removed due to its redundancy and unneeded complexity in the codebase. The primary libraries now used are boost_asio and boost_beast, which were reflected in the changes to the CMakeLists.