Refactor DNS resolver methods and header fields conversion
The resolver's create_query for PTR records has been moved from the source file to the header to align with templating practices. Additionally, the async_receive output in AsyncResolve is no longer captured in a variable as it seemed redundant. In dns_header, a unique_ptr was implemented in tie_dns_header to manage memory dynamically and a new method was introduced