Add NS record parsing test, remove unused test data
In `dns_answer.cpp`, a new test has been added to check the parsing of 'NS' type DNS records. This test ensures correct parsing of the 'NS' type data and cross-checks the output with the expected value. Meanwhile, in `common.cpp`, a block of commented-out test data has been removed as it was no longer being used in any tests.